Testing of the pressure and temperature measurments, pressure and temp are now represented as integers, eg 2007 degrees shall be interpreted  as 20.07 degrees.

+++ Arduino/src/Mainc/SensorDataStruct.h
@@ -56,8 +56,8 @@ struct sensordata {
   /*Accelerations*/
   VectorInt16 acc;
 
-  float temperature;
-  float pressure;
+  int32_t temperature;
+  int32_t pressure;
   float height;
   ps3Controller ps3;
 };
